# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  realNumberDNOClone_internal_models.BulkQueryRequest:
    properties:
      channel:
        type: string
      phoneNumbers:
        items:
          type: string
        type: array
    type: object
  realNumberDNOClone_internal_models.BulkQueryResponse:
    properties:
      hits:
        type: integer
      misses:
        type: integer
      results:
        items:
          $ref: '#/definitions/realNumberDNOClone_internal_models.DNOQueryResponse'
        type: array
      total:
        type: integer
    type: object
...
  realNumberDNOClone_internal_models.CreateUserRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      orgId:
        type: integer
      password:
        type: string
      role:
        type: string
    type: object
...
  realNumberDNOClone_internal_models.DNONumber:
    properties:
      addedByOrgId:
        type: integer
      addedByUserId:
        type: integer
      channel:
        type: string
      createdAt:
        type: string
      dataset:
        type: string
      id:
        type: integer
      investigationId:
        type: string
      numberType:
        type: string
      phoneNumber:
        type: string
      reason:
        type: string
      status:
        type: string
      threatCategory:
        type: string
      updatedAt:
        type: string
    type: object
...
  realNumberDNOClone_internal_models.DNOQueryResponse:
    properties:
      channel:
        type: string
      dataset:
        type: string
      isDno:
        type: boolean
      lastUpdated:
        type: string
      phoneNumber:
        type: string
      status:
        type: string
    type: object
...
  realNumberDNOClone_internal_models.ITGIngestRequest:
    properties:
      channel:
        type: string
      investigationId:
        type: string
      phoneNumber:
        type: string
      threatCategory:
        type: string
    type: object
...
  realNumberDNOClone_internal_models.User:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      orgId:
        type: integer
      role:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: FakeNumber DNO
    url: https://github.com/drewkosta/realNumberDNOClone
  description: |-
    Do Not Originate database API for preventing illegal robocalls and caller ID spoofing.
    Carriers query the DNO database in real-time to block spoofed calls.
  title: FakeNumber DNO API
  version: "1.0"
paths:
  /admin/api-keys:
    delete:
      description: Revoke an organization's API key
      parameters:
      - description: Organization ID
        in: query
        name: orgId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke API key
      tags:
      - Admin
    post:
      description: Generate a new API key for an organization
      parameters:
      - description: Organization ID
        in: query
        name: orgId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              apiKey:
                type: string
              note:
                type: string
              orgId:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Generate API key
      tags:
      - Admin
  /admin/itg-ingest:
    post:
      consumes:
      - application/json
      description: Ingest a phone number from an ITG traceback investigation
      parameters:
      - description: ITG ingest details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/realNumberDNOClone_internal_models.ITGIngestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.DNONumber'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ingest ITG number
      tags:
      - Admin
  /admin/npac-event:
    post:
      consumes:
      - application/json
      description: Process a mock NPAC number porting event
      parameters:
      - description: Porting event details
        in: body
        name: body
        required: true
        schema:
          properties:
            newOwnerOrgId:
              type: integer
            newStatus:
              type: string
            phoneNumber:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: NPAC porting event
      tags:
      - Admin
  /admin/reset-password:
    post:
      consumes:
      - application/json
      description: Admin resets another user's password
      parameters:
      - description: User ID and new password
        in: body
        name: body
        required: true
        schema:
          properties:
            newPassword:
              type: string
            userId:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reset user password
      tags:
      - Admin
  /admin/tss-sync:
    post:
      description: Trigger a mock TSS registry synchronization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              numbersAdded:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: TSS registry sync
      tags:
      - Admin
  /admin/users:
    post:
      consumes:
      - application/json
      description: Create a new user account (admin only)
      parameters:
      - description: User details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/realNumberDNOClone_internal_models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - Admin
...
  /dno/query:
    get:
      consumes:
      - application/json
      description: Check if a phone number is on the Do Not Originate list
      parameters:
      - description: 10-digit phone number
        example: "5551234567"
        in: query
        name: phoneNumber
        required: true
        type: string
      - default: voice
        description: voice or text
        in: query
        name: channel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.DNOQueryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - APIKeyAuth: []
        BearerAuth: []
      summary: Single DNO lookup
      tags:
      - Query
  /dno/query/bulk:
    post:
      consumes:
      - application/json
      description: Check up to 1000 phone numbers against the DNO list in a single
        request
      parameters:
      - description: Phone numbers and channel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/realNumberDNOClone_internal_models.BulkQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.BulkQueryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - APIKeyAuth: []
        BearerAuth: []
      summary: Bulk DNO lookup
      tags:
      - Query
...
securityDefinitions:
  APIKeyAuth:
    description: Organization API key for machine-to-machine access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token (prefix with "Bearer ")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"